# module_I/API.yaml
openapi: 3.0.3
info:
  title: Game Score Manager API
  version: 1.3.10
  description: El proyecto Sistema de Gestión de Puntuaciones es una aplicación web fullstack que permite gestionar y mostrar las puntuaciones (scores) de los jugadores de un videojuego en línea.
  contact:
    email: kevinmorenos@gmail.com
    name: Kevin Moreno
  license:
    name: Apache 2.0
    url: "https://www.apache.org/licenses/LICENSE-2.0"
servers:
  - url:  'http://localhost:3000/api/v1/'
  - url:  'http://172.22.202.139:3000/api/v1'
...
paths:
...
  /users/scores/{userId}:
    get:
      summary: Get user scores
      tags:
        - Scores
      description: List all scores of the player.
      operationId: getUserScores
      parameters:
        - name: userId
          in: path
          required: true
          description: ID of the user whose scores are being retrieved.
          schema:
            type: string
            example: 'uuid'
      responses:
        '200':
          description: User scores retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScoreResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse404'
  /scores/{userId}:
    post:
      summary: Add a new score
      tags:
        - Scores
      description: Add a new score from the video game.
      operationId: addUserScore
      parameters:
        - name: userId
          in: path
          required: true
          description: ID of the user adding the score.
          schema:
            type: string
            example: 'uuid'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoreRequest'
      responses:
        '201':
          description: Score added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScoreResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse400'
...
  /users/admin/scores/{userId}:
    delete:
      summary: Delete a specific score
      tags:
        - Users
      description: Only admin can delete a specific score.
      operationId: deleteUserScore
      parameters:
        - name: userId
          in: path
          required: true
          description: ID of the user whose score is being deleted.
          schema:
            type: string
            example: 'uuid'
      responses:
        '204':
          description: Score deleted successfully
        '404':
          description: User or score not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse404'
components:
  schemas:
...
    ScoreRequest:  
      type: object
      properties:
        score:
          type: integer
          example: 1500
    ScoreResponse:
      type: object
      properties:
        userId:
          type: string
          example: 'uuid'
        score:
          type: integer
          example: 1000
        createdAt:
          type: string
          format: date-time
          example: "2024-10-06T10:15:30Z"
...
    ErrorResponse404:
      type: object
      properties:
        message:
          type: string
          example: "User not found"
        statusCode:
          type: integer
          example: 404
    ErrorResponse400:
      type: object
      properties:
        message:
          type: string
          example: "Bad request"
        statusCode:
          type: integer
          example: 400
...